requirements.txt
setuptools~=56.0.0
requests~=2.25.1
beautifulsoup4~=4.9.3
PyYAML~=5.4.1
numpy~=1.20.3
Jinja2~=3.0.1
APScheduler~=3.7.0
Flask~=2.0.1
gunicorn~=20.1.0
SQLAlchemy~=1.4.19
mysqlclient~=2.0.3
sqlalchemy-pagination~=0.0.2